# Conflicts:
#	examples/ESACCI-LC-L4-LCCS-Map-300m-P1Y-1992-v2.0.7cds/item.json
#	tests/constants.py
#	tests/test_commands.py

+++ tests/test_commands.py
@@ -20,7 +20,7 @@ class CommandsTest(CliTestCase):
 
     def test_create_collection(self) -> None:
         with TemporaryDirectory() as tmp_dir:
-            src_file = os.path.join(constants.SRC_FOLDER, "collection.json")
+            src_file = os.path.join(constants.TRUTH_FOLDER, "collection.json")
             destination = os.path.join(tmp_dir, "collection.json")
 
             result = self.run_command(
@@ -59,18 +59,23 @@ class CommandsTest(CliTestCase):
             id = constants.TEST_FILES[key]["id"]
             path = constants.TEST_FILES[key]["path"]
             with self.subTest(id=id):
-                with TemporaryDirectory() as tmp_dir:
-                    src_data_filename = f"{id}.nc"
-                    stac_filename = f"{id}.json"
-
-                    src_collection = os.path.join(
-                        constants.SRC_FOLDER, "collection.json"
+                src_data_filename = f"{id}.nc"
+                src_data_file = os.path.join(constants.SRC_FOLDER, src_data_filename)
+                if not os.path.exists(src_data_file):
+                    pytest.skip(
+                        f"Data file {src_data_file} not available for test, skipping"
                     )
-                    src_data_file = path
+                    continue
+
+                with TemporaryDirectory() as tmp_dir:
+                    stac_filename = f"{id}.json"
+                    src_collection = os.path.join(
+                        constants.TRUTH_FOLDER, "collection.json"
+                    )
                     dest_data_file = os.path.join(tmp_dir, src_data_filename)
                     shutil.copyfile(src_data_file, dest_data_file)
 
-                    src_stac = os.path.join(constants.SRC_FOLDER, stac_filename)
+                    src_stac = os.path.join(constants.TRUTH_FOLDER, stac_filename)
                     dest_stac = os.path.join(tmp_dir, stac_filename)
 
                     cmd = (
@@ -103,6 +108,11 @@ class CommandsTest(CliTestCase):
                     else:
                         for key in constants.DATA_VARIABLES:
                             del truth_item["assets"][key]
+                        for i in range(len(truth_item["stac_extensions"])):
+                            ext = truth_item["stac_extensions"][i]
+                            if "/raster/" in ext:
+                                del truth_item["stac_extensions"][i]
+                                break
 
                     diff = DeepDiff(
                         item,
